--- lib/view/teams/teamsList.dart
@@ -1,14 +1,0 @@
-import 'package:Runbhumi/widget/widgets.dart';
-import 'package:flutter/material.dart';
-
-class TeamsList extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: buildTitle(context, "Teams"),
-      ),
-      body: Container(),
-    );
-  }
-}
