+++ lib/view/teams/teamsPage.dart
@@ -1,0 +1,26 @@
+import 'package:Runbhumi/widget/widgets.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_icons/flutter_icons.dart';
+
+class TeamsList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: buildTitle(context, "Teams"),
+      ),
+      body: Container(),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          Navigator.pushNamed(context, "/createteam");
+        },
+        child: Icon(Feather.user_plus),
+        foregroundColor: Colors.white,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.all(Radius.circular(20)),
+        ),
+        backgroundColor: Theme.of(context).primaryColor,
+      ),
+    );
+  }
+}
